+++ src/primitives/PieceImage/geometries/parts/bishopBase.ts
@@ -1,0 +1,15 @@
+import type { LathePoints } from "./pointTypes";
+import { Vector2 } from "three";
+
+export const bishopBase: LathePoints = [
+  new Vector2(0, 0),
+  new Vector2(0.09, 0),
+  new Vector2(0.09, 0.03),
+  new Vector2(0.06, 0.03),
+  new Vector2(0.04, 0.09),
+  new Vector2(0.03, 0.15),
+  new Vector2(0.05, 0.15),
+  new Vector2(0.05, 0.16),
+  new Vector2(0.035, 0.16),
+  new Vector2(0, 0.16),
+];
