+++ src/primitives/PieceImage/geometries/parts/index.ts
@@ -1,0 +1,4 @@
+export * from "./bishopBase";
+export * from "./horseNeckAndHead";
+export * from "./rook";
+export * from "./pointTypes";
