+++ src/primitives/PieceImage/geometries/parts/pointTypes.ts
