+++ src/primitives/PieceImage/geometries/rookGeometry.ts
@@ -1,4 +1,4 @@
 import { LatheGeometry } from "three";
-import { rook } from "./shared";
+import { rook } from "./parts";
 
 export const rookGeometry = new LatheGeometry(rook.points, 20);
