--- src/primitives/PieceImage/geometries/shared/index.ts
@@ -1,3 +1,0 @@
-export * from "./parts";
-export * from "./utilities";
-export * from "./types";
