+++ src/primitives/PieceImage/geometries/utilities/index.ts
@@ -1,0 +1,3 @@
+export * from "./getMinOrMaxY";
+export * from "./buildShape";
+export * from "./buildBeveledExtrudedGeometry";
